+++ xivo_confd/plugins/call_filter_fallback/notifier.py
@@ -1,0 +1,21 @@
+# -*- coding: utf-8 -*-
+# Copyright 2018 The Wazo Authors  (see the AUTHORS file)
+# SPDX-License-Identifier: GPL-3.0+
+
+from xivo_bus.resources.call_filter.event import EditCallFilterFallbackEvent
+
+from xivo_confd import bus
+
+
+class CallFilterFallbackNotifier(object):
+
+    def __init__(self, bus):
+        self.bus = bus
+
+    def edited(self, call_filter):
+        event = EditCallFilterFallbackEvent(call_filter.id)
+        self.bus.send_bus_event(event)
+
+
+def build_notifier():
+    return CallFilterFallbackNotifier(bus)